requirements.txt
tensorflow==1.14
pandas==0.25.3
Pillow==6.2.1